# HW2-MarketplaceAPI/catalog-service/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: "Marketplace API"
  version: "1.0.0"
  description: "API для управления товарами и заказами маркетплейса"
servers:
  - url: "http://localhost:8000"
    description: "Local development"
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: "Обновление access token"
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: "Новые токены"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/RefreshTokenInvalid'
...
  /orders:
    post:
      operationId: create_order
      summary: "Создание заказа"
      description: "Роли: USER, ADMIN. SELLER — запрещено"
      tags: [orders]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: "Заказ создан"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '404':
          $ref: '#/components/responses/ProductNotFound'
        '409':
          description: "Конфликт"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                product_inactive:
                  value:
                    error_code: PRODUCT_INACTIVE
                    message: "Товар неактивен"
                    details: null
                order_has_active:
                  value:
                    error_code: ORDER_HAS_ACTIVE
                    message: "У пользователя уже есть активный заказ"
                    details: null
                insufficient_stock:
                  value:
                    error_code: INSUFFICIENT_STOCK
                    message: "Недостаточно товара на складе"
                    details:
                      items:
                        - product_id: "uuid"
                          requested: 10
                          available: 5
                invalid_state:
                  value:
                    error_code: INVALID_STATE_TRANSITION
                    message: "Недопустимый переход состояния"
                    details: null
        '422':
          description: "Ошибка промокода"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                promo_invalid:
                  value:
                    error_code: PROMO_CODE_INVALID
                    message: "Промокод недействителен"
                    details: null
                promo_min_amount:
                  value:
                    error_code: PROMO_CODE_MIN_AMOUNT
                    message: "Сумма заказа ниже минимальной для промокода"
                    details: null
        '429':
          $ref: '#/components/responses/OrderLimitExceeded'
  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: get_order
      summary: "Получение заказа по ID"
      description: "Роли: USER — только свои, ADMIN — любые. SELLER — запрещено"
      tags: [orders]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: "Заказ найден"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/OrderOwnershipViolation'
        '404':
          $ref: '#/components/responses/OrderNotFound'

    put:
      operationId: update_order
      summary: "Обновление заказа"
      description: "Роли: USER — только свои, ADMIN — любые. SELLER — запрещено"
      tags: [orders]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: "Заказ обновлён"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/OrderOwnershipViolation'
        '404':
          $ref: '#/components/responses/ProductNotFound'
        '409':
          description: "Конфликт"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/OrderLimitExceeded'
  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: cancel_order
      summary: "Отмена заказа"
      description: "Роли: USER — только свои, ADMIN — любые. SELLER — запрещено"
      tags: [orders]
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: "Заказ отменён"
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/OrderOwnershipViolation'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/InvalidStateTransition'
  /promo-codes:
    post:
      operationId: create_promo_code
      summary: "Создание промокода"
      description: "Роли: SELLER, ADMIN"
      tags: [promo-codes]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: "Промокод создан"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/TokenInvalid'
        '403':
          $ref: '#/components/responses/AccessDenied'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ProductNotFound:
      description: "Товар не найден"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: PRODUCT_NOT_FOUND
            message: "Товар не найден по ID"
            details: null
    OrderNotFound:
      description: "Заказ не найден"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ORDER_NOT_FOUND
            message: "Заказ не найден по ID"
            details: null
    ValidationError:
      description: "Ошибка валидации"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: "Ошибка валидации входных данных"
            details:
              errors:
                - field: "name"
                  message: "Field required"
    TokenInvalid:
      description: "Невалидный или просроченный токен"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired:
              value:
                error_code: TOKEN_EXPIRED
                message: "Токен просрочен"
                details: null
            invalid:
              value:
                error_code: TOKEN_INVALID
                message: "Невалидный токен"
                details: null
    RefreshTokenInvalid:
      description: "Невалидный refresh token"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: REFRESH_TOKEN_INVALID
            message: "Невалидный или просроченный refresh token"
            details: null
    AccessDenied:
      description: "Недостаточно прав"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ACCESS_DENIED
            message: "Недостаточно прав доступа"
            details: null
    OrderOwnershipViolation:
      description: "Заказ принадлежит другому пользователю"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ORDER_OWNERSHIP_VIOLATION
            message: "Заказ принадлежит другому пользователю"
            details: null
    OrderLimitExceeded:
      description: "Превышен лимит частоты операций"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ORDER_LIMIT_EXCEEDED
            message: "Превышен лимит частоты создания/обновления заказа"
            details: null
    InvalidStateTransition:
      description: "Недопустимый переход состояния"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: INVALID_STATE_TRANSITION
            message: "Недопустимый переход состояния заказа"
            details: null
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
    OrderItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999
    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'
        promo_code:
          type: string
          nullable: true
          pattern: "^[A-Z0-9_]{4,20}$"
    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemCreate'
    OrderItemResponse:
      type: object
      required: [product_id, quantity, price_at_order]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal
    OrderResponse:
      type: object
      required: [id, user_id, status, total_amount, discount_amount, items, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]{4,20}$"
        discount_type:
          type: string
          enum: [PERCENTAGE, FIXED_AMOUNT]
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
          default: true
    PromoCodeResponse:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          type: string
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true